use patched version of nixStable
The rust cli wraps the bash cli, so should use the same version of
nix. We chose to copy the patch files rather than reference them
from the bash cli repository because this will soon be the ultimate
canonical home of those patches and we can save time overall by
going straight to the end state with this patch. If for some reason
we need to update the patch over the next month (unlikely) then we
can update the patches in both places.