feat(runix): Additional flags for eval and flake commands (#40)
* feat(runix): add `--expr` flag

Add the shared `--expr` Nix flag and the appropriate command grouping
SourceArgs which includes it.

* feat(runix): add `--no-write-lock-file` flag

Adds the `--no-write-lock-file` flag to FlakeArgs

* refactor(runix): expose installable fields publically

Marks the fields on `Installable` as public so that it can be
constructed directly externally

* feat(runix): Add `eval` Command

Adds the `nix eval` command. Also add EvalArgs for `nix eval`
specific arguments which includes the `--apply` argument, and implement `ToArgs` for `InstallableArg` which `Eval` uses.