api: make `installable` argument a flag
As a positional it is undecidable if the first positional is an installable
or a nix argument.
This is unless we can parse and validate the installable.
The better alternative is to separate nix flags using `--`
but that syntax is incompatible with `flox.sh` which most/all commands are passed down to.