feat: Add environment ref support to CLI
* feat(installables): abstract support for `-e` as an installable
Update environment ref logic, migrate our data from `flox-preview` to `flox`, add PosOrEnv enum for allowing `-e [env]` in place of a positional installable, and enable PosOrEnv for `bundle`

The argument handling logic is not very clean and is incomplete, but makes this a working demonstration

* feat(sdk): theoretically working installable generation from EnvironmentRef
Untested but follows the rough spec of how this should work, also includes a few tweaks to the existing EnvironmentRef logic

* refactor(sdk): Improve Git provider and environment ref implementation
Clean up various parts of the environment ref implementation and use our GitProvider system which now supports what it requires

* feat: add EnvironmentRef struct
EnvironmentRef parses user input into either a reference to a project
(e.g. for ./pkgs/default) or a named environment (e.g. for
mkenigs/default)