feat(env-ref): use new installable-based project ref format and add resolver
Since they are now installables, we can re-use our installable resolution logic, and copy how we prompt the user to choose (with some environment-specific changes). This also makes it possible to be more ambiguous with your environment choice and get prompted when required (when we default and when we prompt can be decided later).