feat(cli): no longer force first component of key as prefix
Previously if more than 1 key component was provided, the first one would always be used as the prefix (before the system). This change removes setting the prefix from the Rust side, and the resolver now attempts to use it both as a prefix and also appended to the default prefixes.

Resolves #295