fix: deprecate gitCheckout()
We are no longer checking out branches of the floxmeta repository
in situ and are instead performing all manipulations using a bare
repository. With that we no longer need to call gitCheckout() from
metaGit(), after which that function is no longer used at all.

Also pared an invocation of metaGit() that was never being performed
on account of $_cline never being defined in that scope.