fix: regression regarding `flox destroy` branch removal
The move to bare repositories broke the ability to checkout the
floxmain branch, and that in turn broke the logic in `flox destroy`
to delete environment branches. This patch fixes the issue by
removing the requirement to check out the "floxmain" branch prior
to destroying the environment branch. In previous versions we
performed this checkout as a way of guaranteeing we would never
attempt to delete the currently-checked-out branch.