fix(gh): allow missing oauth_token (#64)
`gh` introduced support for storing credentials in an encrypted keyring
about 3 months ago ([1])

If stored in a keyring the `oauth_token` field in the hosts.yml file
will remain empty. We previously relied on that field to be present,
drawing only from observations due to the lack of documentation on the
schema of `hosts.yml`.

Since reading the hosts.yml file is no reliable source for the user's
access token anymore, we should either employ

    gh auth token --hostname <host>

or REVISIT THE USE OF `gh` ALTOGETHER.

[1]
https://github.com/cli/cli/commit/df83dc2d58f242fc368fa62656d410e5fd820f6e



fixes #44
fixes #22