fix: don't invoke nix build with --impure flag when publishing
The placement of --impure within the floxBuild() function meant that
it also applied to nix builds happening by way of flox publish
invocations. This patch moves the flag to the [1] place where
we are calling floxBuild to build flox environments, which do
require the use of --impure.

Note that flox build as called from the top-level flox-bash/flox.sh
is handled by rust, so that use of the floxBuild() function is not
relevant.