fix: fix catalog parsing for flake upgrades
This commit partially reverts commit 7d16e9bb which broke the ability
to detect available catalog entry upgrades and fixes the ability to
detect available flake upgrades. It is being rolled out as a hotfix
to flox-prerelease as this is currently breaking the functionality
of `flox upgrade` in the majority of cases.