fix: catalog parsing for flake upgrades (#91)
This commit partially reverts commit 7d16e9bb which broke the ability to
detect available catalog entry upgrades and fixes the ability to detect
available flake upgrades. It is being rolled out as a hotfix to
flox-prerelease as this is currently breaking the functionality of `flox
upgrade` in the majority of cases.

## Checks

- [x] All tests pass.
- [x] All commits in this Pull Request are
[signed](https://docs.github.com/en/authentication/managing-commit-signature-verification/signing-commits)
and Verified by Github.
- [x] I have accepted the flox [Contributor License
Agreement](../blob/main/.github/CLA.md) by adding my Git/Github details
in a row at the end of the
[CONTRIBUTORS.csv](../blob/main/.github/CONTRIBUTORS.csv) file by way of
this pull request or one done previously.

## Release Notes

* Fixed bug introduced with #74 which affected the ability to upgrade
packages (both flakes and catalog packages) with `flox upgrade`.