trace(): remove unnecessary quotes from echo invocation
The trace() function was echoing a string containing an unquoted
embedded pprint() invocation, and the resulting string was being
globbed to the empty string on account of `shopt -s extglob`.
This problem was evident when running `flox --debug --debug` as
in the following example:

    $ flox --debug --debug destroy -e foo |& grep "Are you"
    + boolPrompt 'Are you sure?' no
    + trace 'Are you sure?' no
    trace:/nix/store/8x517fn3ckdgj4p013hnbyl91wsg9g0f-flox-bash-0.2.3-r532/lib/commands/environment.sh:1365 boolPrompt( 'Are you no )
    + local 'prompt=Are you sure?'

This patch removes a pair of quotes so that the output of pprint()
is quoted, thus preventing the glob'ing from removing arguments.
With this patch in place the above example shows:

    $ flox --debug --debug destroy -e foo |& grep "Are you"
    + boolPrompt 'Are you sure?' no
    + trace 'Are you sure?' no
    trace:/nix/store/fgz91ycj7qzz76d9snwp4vdg4c5g6kla-flox-bash-0.2.3-dirty/lib/commands/environment.sh:1365 boolPrompt( 'Are you sure?' no )
    + local 'prompt=Are you sure?'