fix: compare to correct branch, synchronize local and upstream before push
We previously compared to HEAD of the upstream default branch regardless
what branch we were pushing from. This commit ensures that we're comparing
branches of the same "ref" name, and it also prompts to push changes upstream
in the event that the local clone is out of date with respect to upstream.