fix: force push floxmain when creating repository on floxhub
The floxhub migration process starts with a fetch, which creates a new
repo with a floxmain branch created from github:flox/floxmeta-template.
If a user's floxmain branch was created long ago they won't have a common
commit history with the template, so we must force push (this one branch
only) to bring them into sync.

This approach does introduce the possibility that users will migrate
different machines and end up with different contents of their floxmain
branch in the process, but that's really just a symptom of them having
been out of sync in the first place.