feat: use flox customed compilation of nixpkgs#gh
The github version of `gh` demands very powerful scopes, but all we
need with flox is a token capable of authenticating a user's GitHub
account. With this patch we are modifying the `gh` package to remove
all scopes from the token being acquired so that it can only be used
for authentication.