feat: use `gh auth token --git-credential-helper`
The `gh auth token` command very nearly provided everything necessary
for use as a git credential helper, but was lacking the username field.
This commit updates our build of the `gh` command to add a new
`gh auth token --git-credential-helper` flag which emits the username
and the password in the standard git credential helper format.