fix: more tweaks to flox-gh.patch
Found a few more instances where we can tune the `gh` output to
redact all mentions of "gh" itself. This is clearly a work in progress
until such time as we rewrite this as a selection of the functionality
we need rather than a redaction of functionality we don't need.