fix: adapt floxmetaHelperGit() for use with different URLs
Applying a fixed git configuration from within floxmetaHelperGit()
while we still have floxmeta repositories sourced on github.com was
introducing a change from previous flox versions which utilized the
user's own gitconfig for accessing github.com in a variety of ways.

This patch changes that, causing floxmetaHelperGit() to invoke git
differently based on the remote URL it is talking to. Notably, this
pulls in the specification of the floxHub credential helper as a
command-line argument, which allows us to deprecate all the logic
for maintaining a flox-specific gitconfig file.