fix: disregard gh environment variables
Turns out that github:cli/cli pulls in github:cli/go-gh as a dependency,
and it's in that module that the GITHUB_TOKEN environment variable is
considered to trump the local hosts.yml as managed by github:cli/cli.

Rather than pull a second module to be patched, it's easier to clean
the environment of all problematic variables in the flox-gh wrapper.