chore: add Justfile for development (#219)
## Proposed Changes

<!-- Describe the changes proposed in this pull request. -->
<!-- Please provide links to any issue(s) which are expected to be
resolved. -->
This PR adds `just` to the `rust-env` environment and a Justfile for
local development. The Justfile has commands for running the test suite
and makes it easier to run subsets of the `bats` test suite
specifically.

Example:
```
$ just bats-tests ./tests/activate.bats
```

## Current Behavior

<!-- Describe the current behavior before applying this pull request.
-->
Currently if you want to run a specific test file you would have to:
```
$ cargo build
$ flox run '.#flox-tests' -- -- --flox target/debug/build ./tests/activate.bats
```

## Checks

<!-- Please confirm the following: -->

- [x] All tests pass.
- [x] All commits in this Pull Request are
[signed](https://docs.github.com/en/authentication/managing-commit-signature-verification/signing-commits)
and Verified by Github.
- [x] I have accepted the flox [Contributor License
Agreement](../blob/main/.github/CLA.md) by adding my Git/Github details
in a row at the end of the
[CONTRIBUTORS.csv](../blob/main/.github/CONTRIBUTORS.csv) file by way of
this pull request or one done previously.


## Release Notes

<!-- Describe any user facing changes. Use "N/A" if not applicable. -->
N/A

<!-- Many thanks! -->