Revert "fix: use HTTP/1.1 for git operations"
This reverts commit 7d88fa251b820fbc0319a92a4257391320e1e901.

Found https://github.com/curl/curl/issues/11353 which confirms that
the curl "stream reset" error is fixed with curl version 8.2.1, which
as it happens is available on the staging stability which will roll
over to stable on Saturday. With that Yannik and I agreed to consign
this bit of knowledge to git history and avoid cluttering the code
with workarounds.