fix: --version flag usage
Some internal docs for `--verbose` were leaking into the usage output of flox.
This PR removes an indirection used for the "fake" `--version` flag.

As a reminder; `--version` is handled separately from the main argument handling,
so that it any flox command that contains `--version` will print the version
and abort early in proper POSIX fashion.