fix: move `delete_symlinks` to `Environment trait`
`delete_symlinks` is used for garbage collection.
Our current approach leaves it to the environment to delete relevant gcroot symlinks
and then calls `nix store gc` as a sort of catch all if any symlink was deleted.
This fix only makes sure that `delete_symlinks` works on generic `Environment` impls.
Whether _this_ approach should be maintained or adapted shall be part of a separate discussion.