fix(regresssion): remove `ident` from `PathEnvrionment::open`
There is at most one environment in any `.flox/` dir.
The `ident` argument was used to choose one of N environments which is now obsolete.
We removed code for reading `.flox` from the cli crate
so logic to determine the name of the env is now contained within the `open` method.