fix: pin upstream gh version at 2.32.1
We're only expecting to carry the `gh` dependency for a few more weeks
so we can pin it at version 2.32.1 for that period of time. We do this
by committing a modified version of the full Nix expression as found
in the previous stable nixpkgs because just overriding the src and
vendor checksums does not yield a working build.