feat: add 'flox show' command (#287)
## Proposed Changes

<!-- Describe the changes proposed in this pull request. -->
<!-- Please provide links to any issue(s) which are expected to be
resolved. -->
This adds a bare bones feature-flagged `flox show` command. The output
is currently just a list of the search results from the provided package
name, but the command is aware of disambiguated packages. Properly
formatted output is coming in later PRs.

Closes flox/product#356.

## Current Behavior

<!-- Describe the current behavior before applying this pull request.
-->

N/A


## Checks

<!-- Please confirm the following: -->

- [x] All tests pass.
- [x] All commits in this Pull Request are
[signed](https://docs.github.com/en/authentication/managing-commit-signature-verification/signing-commits)
and Verified by Github.
- [x] I have accepted the flox [Contributor License
Agreement](../blob/main/.github/CLA.md) by adding my Git/Github details
in a row at the end of the
[CONTRIBUTORS.csv](../blob/main/.github/CONTRIBUTORS.csv) file by way of
this pull request or one done previously.


## Release Notes

<!-- Describe any user facing changes. Use "N/A" if not applicable. -->
N/A

<!-- Many thanks! -->