feat: format flox show output (#327)
## Proposed Changes

<!-- Describe the changes proposed in this pull request. -->
<!-- Please provide links to any issue(s) which are expected to be
resolved. -->
This properly formats the output of the feature-flagged `flox show`
command:

```
$ flox show hello
hello - A program that produces a familiar, friendly greeting
    hello - hello@2.12.1

$ flox show hello --all
hello - A program that produces a familiar, friendly greeting
    hello - hello@2.12.1, hello@2.12, hello@2.10
```

## Current Behavior

<!-- Describe the current behavior before applying this pull request.
-->
The current output is a crude list of `name - description` lines.

## Checks

<!-- Please confirm the following: -->

- [x] All tests pass.
- [x] All commits in this Pull Request are
[signed](https://docs.github.com/en/authentication/managing-commit-signature-verification/signing-commits)
and Verified by Github.
- [x] I have accepted the flox [Contributor License
Agreement](../blob/main/.github/CLA.md) by adding my Git/Github details
in a row at the end of the
[CONTRIBUTORS.csv](../blob/main/.github/CONTRIBUTORS.csv) file by way of
this pull request or one done previously.


## Release Notes

<!-- Describe any user facing changes. Use "N/A" if not applicable. -->
N/A

<!-- Many thanks! -->

---------

Co-authored-by: Alex Ameen <alex.ameen.tx@gmail.com>