feat: edit manifest with 'flox uninstall' (#340)
## Proposed Changes

<!-- Describe the changes proposed in this pull request. -->
<!-- Please provide links to any issue(s) which are expected to be
resolved. -->
This PR edits `manifest.toml` with `flox uninstall`.

Given a `manifest.toml` that looks like
```toml
[install]
foo = {}
bar = {}
```
and the uninstallation command you'll see the following output
```
$ flox uninstall foo bar
🗑️ 'foo' uninstalled from environment
🗑️ 'bar' uninstalled from environment
```

If _any_ of the packages requested for uninstallation is missing from
the manifest the whole operation is aborted and the manifest is
unchanged. This is to pave the way for transactional operations.


## Release Notes

<!-- Describe any user facing changes. Use "N/A" if not applicable. -->
`flox uninstall` removes packages from `manifest.toml`

<!-- Many thanks! -->