refactor: deprecate dead legacy code (#351)
Moves most legacy code and other dead code paths to the git history
graveyard.

Commits of this PR aim to be self contained only removing individual
features.
The hope is that this will make it easier to resurrect these concepts if
project requirements call for it or we are looking for reference
implementations as we redesign features.