feat: auth device flow (#341)
## Proposed Changes

<!-- Describe the changes proposed in this pull request. -->
<!-- Please provide links to any issue(s) which are expected to be
resolved. -->
This change implements a portion of the device flow auth login for
github in rust code native to the flox cli as a temporary cli command
called `auth2` once the entire device flow is implemented `auth2` will
be switched to become `flox auth` command sub-set. This PR implements
the `flox auth2 login` but does not store and use the token **because
this is a partial task toward the total completion**

**Please see https://github.com/flox/product/issues/521 for the specific
acceptance criteria of this issue**

## Release Notes

<!-- Describe any user facing changes. Use "N/A" if not applicable. -->
None for now see acceptance criteria at
https://github.com/flox/product/issues/521#issue-1945568642

<!-- Many thanks! -->

---------

Co-authored-by: Yannik Sander <7040031+ysndr@users.noreply.github.com>
Co-authored-by: Yannik Sander <yannik@floxdev.com>