feat: activate builds environments (#372)
## Proposed Changes

This PR builds environments when `flox activate` is called. The
environment is built with a new executable called `build-env`, which
currently is just a shell script that calls `nix build` on another new
file: `env-from-lockfile.nix` .

The `activate` command currently builds a hard-coded lockfile.
Generating a lockfile from the manifest is coming in a later PR.

This PR also disables `clippy` as a precommit hook because it's super
annoying.

## Release Notes

<!-- Describe any user facing changes. Use "N/A" if not applicable. -->
`flox activate` now builds an environment.

<!-- Many thanks! -->