feat: display progress message during activate (#391)
## Proposed Changes

<!-- Describe the changes proposed in this pull request. -->
<!-- Please provide links to any issue(s) which are expected to be
resolved. -->
Displays a message during a build triggered by `flox activate` so that a
user has some indication that something is going on and the program
isn't stalled.

The message is written to `stderr` and then dismissed when the build is
complete.

## Release Notes

<!-- Describe any user facing changes. Use "N/A" if not applicable. -->
A progress message is displayed while building an environment as a
result of calling `flox activate`.


<!-- Many thanks! -->