fix: warm up pkgdb to avoid timeout
pkgdb must download and evaulate nixpkgs, which is time intensive. At
times this causes expect to time out.

Instead of bumping the timeout, factor out operations that we never want
to timeout outside of the expect test. Since expect must have a timeout,
the timeout will be easier to set accurately if we move actions that
should never timeout outside of expect. This may also make caching
easier in the future.