fix: don't allow crane to link inherited artifacts
The 2023-08-07 revision of crane changed the behavior to create symlinks
to dependency .rlib and .rmeta files in the temporary build "target"
directory. This creates a problem for Nix builds as these symlinks
will point to immutable files in the store.

Set the new "doNotLinkInheritedArtifacts" variable to ensure this
does not cause problems going forward.

See https://crane.dev/CHANGELOG.html