feat: resolve owner to push for using github token (#419)
Uses the GitHubClient introduced in #410 to determine the github
username associated with the login token we have already stored.

The username is parsed as an EnvironmentOwner so _some_ users with may
be unable to use floxhub at all using plain github handles. This is
understood to change during the transition to FloxHub users.

closes #418