feat: format errors from pgkdb (#446)
## Proposed Changes

<!-- Describe the changes proposed in this pull request. -->
<!-- Please provide links to any issue(s) which are expected to be
resolved. -->
This adds a custom std::error::Error implementation for `PkgDbError`
such that the JSON we receive from `pkgdb` is printed as if it we any
other encountered on the Rust side.

Example:
```
$ flox search hello
ERROR: search encountered an error

    Caused by:
        0: invalid manifest file
        1: unrecognized manifest field `options.doesnt_exist'.
```

Previously the JSON errors received from `pkgdb` would simply be printed
in their raw form.

## Release Notes

<!-- Describe any user facing changes. Use "N/A" if not applicable. -->
Users will now see friendly error messages when there's an error with
their manifest or when an error is encountered during a search.

<!-- Many thanks! -->