test: add integ tests for manifest presence (#453)
## Proposed Changes

<!-- Describe the changes proposed in this pull request. -->
<!-- Please provide links to any issue(s) which are expected to be
resolved. -->
This adds a couple of tests for `search`/`show` that ensure these
commands fall back to the global manifest when outside of a project.

As part of making the tests pass some helper functions were added that
aid in discovering a `.flox` directory.

This is the final PR required for closing flox/product#551.

## Release Notes

<!-- Describe any user facing changes. Use "N/A" if not applicable. -->
N/A

<!-- Many thanks! -->