feat: add a `--force` flag for flox (`push|pull`) (#449)
## Proposed Changes

This pr adds the use of a `--force` flag on `flox push` and `flox pull`
commands.
Since modifying commands for `ManagedEnvironment`s are not yet
implemented, we're right now unable to test `flox pull --force` in any
realistic way easily -- deferred to after more of `ManagedEnvironment`
is implemented.


## Release Notes

N/A

---------

Co-authored-by: Yannik Sander <7040031+ysndr@users.noreply.github.com>
Co-authored-by: Yannik Sander <yannik@floxdev.com>