fix: replace transaction .flox by copying (#467)
`fs::rename` requires both source and destination paths to be on the
same file system. Since the project may be on any mount point on the
system, it can not be guaranteed that restoring the .flox copy under
transaction by renaming will succeed.

This change will finish the transaction by _copying_ back the modified
.flox rather than trying to rename it.