feat: make environment detection consistent (#558)
Previously, search/show searched upwards for a .flox, while environment
commands did not. Environment commands used the last activated
environment, but search/show did not. Instead, all commands (except
activate) should both search up and use the last activated environment.

- Introduces `UninitializedEnvironment` to represent an environment with
a valid env.json but that may not be fully initialized. This can be
used, for example, to prompt users to pick between environments without
the slowdown of having to initialize both of them first.
- `find_dot_flox` now returns an `UninitializedEnvironment`.
- `detect_environment` now uses `find_dot_flox` to search upwards.
- `manifest_and_lockfile` was moved to the flox crate, and it uses
`detect_environment` to check for the last activated environment.
- Tests for `find_dot_flox` were disabled, as `find_dot_flox` behavior
will be changed in a follow-up PR to stop at Git boundaries.

---------

Co-authored-by: Yannik Sander <7040031+ysndr@users.noreply.github.com>
Co-authored-by: Yannik Sander <yannik@floxdev.com>