feat: flox upgrade
`flox upgrade` tries to resolve packages in a newer input. Upgrades can
be specified using iid or group name. "toplevel" is introduced as the
group name for packages without a group. `upgrade` will fail if an iid
for a package that is part of a group with other packages is specified.

Code notes:
- In pkgdb, groups are now stored as an unordered_map instead of a
  vector using group name as key, since "toplevel" can be used as the
  key for the toplevel group.
- A bug in `tryResolveGroup` was fixed where packages that should have
  been upgraded were being resolved in the old input.