feat: flox update, introduce global lock (#601)
- Introduce a global manifest lockfile.
- Add an update subcommand that either:
  - Updates an environments inputs.
- Updates inputs in the global manifest, storing the lock in the newly
introduced global lockfile.

Code notes:
- The `pkgdb_errors` module was renamed to `pkgdb`, and a `call_pkgdb`
function was factored out along with some errors.
- The manifest argument to all pkgdb subcommands was made optional to
allow them to operate on the global manifest. Commands will still fail
unless an empty manifest is explicitly set in the implementation of the
command, which is the approach taken by search, and by update in this
commit.
- `transact_with_lockfile_contents` was added to `CoreEnvironment`

---------

Co-authored-by: Alex Ameen <alex.ameen.tx@gmail.com>
Co-authored-by: Yannik Sander <7040031+ysndr@users.noreply.github.com>