feat: add database locking
Several process may all be started at once that interact with the
database. Only one of them will be able to write to the database,
and the others will block until that database operation is done.

This also adds a `pkgdb fingerprint` command which will return the
fingerprint of the flake reference passed on the command line. This
flake reference will be locked before returning the fingerprint.
This command is only meant for testing purposes.