feat: add database locking (#732)
## Proposed Changes

<!-- Describe the changes proposed in this pull request. -->
<!-- Please provide links to any issue(s) which are expected to be
resolved. -->
Adds simple database locking so that more than one search process
launched shortly after one another don't step on each other.

Note that this isn't bulletproof and you'll notice in the tests that the
processes aren't launched at exactly the same time. Instead, the
processes are launched with a slight delay between them. This is good
enough because the processes will *most likely* never be launched at
**exactly** the same time in any real use case.

## Release Notes

<!-- Describe any user facing changes. Use "N/A" if not applicable. -->
N/A

<!-- Many thanks! -->