feat: inject current system during init (#749)
## Proposed Changes

<!-- Describe the changes proposed in this pull request. -->
<!-- Please provide links to any issue(s) which are expected to be
resolved. -->
Adds the `options.systems` field during init populated with the current
system.

`m4` is used to inject the system into the template instead of
`toml_edit` because `toml_edit` injects the `options` table between the
`hook` table and the commented out script below it, so a user that
uncomments the script hook will encounter an error.

## Release Notes

<!-- Describe any user facing changes. Use "N/A" if not applicable. -->
Users that create an environment will now explicitly have their sytem
type added to the manifest.

<!-- Many thanks! -->