fix: remove async telemetry event handling (#766)
The telemetry sending used to happen in an async context on a separate
thread.
Events were sent to the telemetry thread using mpsc and handled
concurrently with the main thread.
Both threads were using a handle to the same global `tokio` runtime.
When the execution of the main thread completes, the async runtime is
dropped.
Any task still on the runtime is aborted.
As telemetry buffering **and transfer** used the same runtime on a
parallel thread,
we observed frequent aborts of the telemetry transfer
(https://github.com/flox/flox/issues/432).

This PR drops the concurrent telemetry thread in favour of immediate
writes to the metrics buffer
and _occasional_ transfer.

* writing on the main thread _will_ incur a small amount of time,
  we hope the delay will be non-noticeable in practice
* every two hours the buffer is flushed to the upstream telemetry
backend,
  this may be a longer delay.
We should monitor whether this is noticeable and needs further
refinement.
* since we write on the main thread, `panic` should be avoided at all
cost.