feat: use global lockfile for all subcommands (#750)
- Use the global lockfile for search/show when not using an
environment's lockfile, creating it if it does not yet exist
- Use the global lockfile for `CoreEnvironment::lock`, creating it if it
does not yet exist. This means that when first installing to an
environment, the environment will be locked to the version of nixpkgs in
the global lock, rather than whatever is latest. If this is not desired,
users can run `flox update --global && flox install` instead of just
`flox install`
- factor out `pkgdb_update` and `update_global_manifest` to allow the
above