fix: Precompute default channels instead of parsing every time
Every invocation of flox loads the default channels, which requires
shelling out to pkgdb to parse each one. At 40ms per pkgdb call, that
adds up to significant latency for otherwise fast commands. Particularly
those that don't need the channels in the first place.

Since these are hardcoded and trivial, we can simply precompute the
parsed FlakeRef representation rather than relying on pkgdb to do it for
us.

In testing, this reduced a `flox activate` from ~200ms to ~50ms.