fix: add test to test failure with absence of LD_AUDIT
My previous attempt to use libgif.so as a library that would be unlikely
to be installed on the base O/S failed when I found it's still installed
by default on fedora.

So ... the best solution would be to create a bespoke custom library
and install it as a package to the env, but as we don't have support
for building custom packages or installing to an environment by
/nix/store path I instead went in search of a shared library that would
be guaranteed not to be found on a system. Ultimately I realized that
libnixmain.so from the Nix package itself could be a possible candidate,
and created a simple test program to demonstrate how LD_AUDIT can be
used to find (or _not_ find if absent) the necessary libraries.