fix: updated comments, fixed forced use of LD_LIBRARY_PATH for glibc
(re)Discovered that the version of the ld interpreter must match the
glibc version in order for it to work reliably, and updated the test
accordingly to ensure that we explicitly set the interpreter before
exercising LD_LIBRARY_PATH to test the floxenv-provided version of
glibc. Also updated the comments to reflect recent changes to this
test.