fix: set Nix env vars with non-interactive activation
The interactive activation of a flox environment sets a number of
Nix and flox-specific environment variables that are important for
the operation of Nix-packaged software in general, and we observed
that these same environment variables are not getting set for the
non-interactive activations (e.g. as is the case when one adds
'eval "$(flox activate)"' to a "dotfile").

This patch sources the same default_nix_subprocess_env() function
that sets these variables in the interactive case when generating
the activation script. Notably it adds a parameter to the function
to control whether it returns an environment map with all variables
set by default, or just containing those variables not already
present in the enviroment.

Closes #817