feat: bump to nixpkgs release 23.11 (#828)
Bumps several things to nixpkgs release 23.11:
- The ref of nixpkgs we use to build flox.
- The ref of nixpkgs we use by default to provide packages for the CLI.
- The revision we use in our tests.

An overlay was added to continue using bear from 23.05, because it is
broken on darwin in 23.11.

The bulk of these changes are changes to the tests. We use 3 different
revs across all of our test suites:
- One from release 23.11
- One from release 23.05
- A revision old enough to provide glibc 2.34

A non-exhaustive list of some of the changes that are non-obvious:
- Introduce NIXPKGS_REV_OLD in setup_suite instead of deducing OTHER_REV
  from lockfiles.
- Drop propagated packages test because it was already skipped.
- Use linux instead of zfs-kernel for prefer pre-release test because
  zfs-kernel no longer has a prerelease.
- Remove some unused test files in pkgdb.

---------

Co-authored-by: Tom Bereknyei <tomberek@gmail.com>