fix: match r and r/w db timeouts (#865)
## Proposed Changes

<!-- Describe the changes proposed in this pull request. -->
<!-- Please provide links to any issue(s) which are expected to be
resolved. -->
The timeouts while waiting for a database scrape were different while
waiting to create a read/write connection vs. a read-only connection.
The read-only timeout was roughly 20 seconds, whereas the read/write
timeout is a little over 4 minutes. This PR makes the read-only timeout
the same as the read/write timeout (~4 minutes).

## Release Notes

<!-- Describe any user facing changes. Use "N/A" if not applicable. -->
N/A

<!-- Many thanks! -->