feat: deduplicate multi-system search results
Currently, search results are deduplicated in Rust, which means:
- Search limit is not implemented correctly, because results are
  deduplicated after the limit is applied rather than before.
- Deduplication is not applied everywhere it should be: DidYouMean
  doesn't deduplicate results for InstallSuggestions.

Instead, use a `GROUP BY relPath` in pkgdb to deduplicate results. While
this violates the separation of knowledge between search/install and
pkgdb, it seems like the best way to get deduplication done for now. If
we want to restore the separation of knowledge, we could expose fields
that toggle `GROUP BY` statements.

Additionally, clean up how search results are handled in Rust. Ensure
that all search results are disambiguated and displayed correctly by
only printing search results with the DisplayItem::fmt method.