fix: access to class method in static context
`GitHubInputScheme::inputFromAttrs` calls `this->type()` which resolved `this` to the calling instance of `FloxFlakeScheme` (?).
I'm not exactly sure why it would do that, but accessing an instance method in a static method seems odd either way.

For reference nix is _also_ calling the `inputFrom*` methods on an _instance_ of the schema:
* https://github.com/NixOS/nix/blob/2.17-maintenance/src/libfetchers/fetchers.cc#L36
* https://github.com/NixOS/nix/blob/2.17-maintenance/src/libfetchers/fetchers.cc#L51