fix: do not pull managed env into a tempdir (#888)
Environments were pulled into a temp directory which would later be
renamed to `.flox`. We did this to avoid populating the floxmeta branch
`owner/name.encode(path to .flox)` in case building fails or the user
aborts the fixup. The branch would be adjusted when the environment is
opened the next time by the `ensure_branch` routine.

Unfortunately, we would be missing the `out_link` linked to the tempdir,
which causes `flox activate` (or any other activity requiring the
activation path) to issue a rebuild of the environment.

Addressing the original concern:
(1) `ensure_branch` is already resetting the env branch if there is a
race condition

<https://github.com/flox/flox/blob/8a69e2d2ab41c1c659dbbb99103ce31bdb3c48fc/cli/flox-rust-sdk/src/models/environment/managed_environment.rs#L684-L699>
(2) if the pull is aborted we could `delete` the environment explicitly
(ctrl+c cases would still need to be covered by (1)

fixes #885