fix: reduce unnecessary rebuilds of remote and managed environments
Pulling a managed environment _always_ causes a rebuild even if no changes were pulled.
That is because each pull would (re) write the generation lockfile at `.flox/env.lock`.
Since rebuilding depends on the mtime of that file, we rebuilt unnecessarily often.

The solution for now is to skip writing the file iff
(1) the file exists
(2) the contents are unchanged