feat: protect active environment state until push succeded
The inner managed environment of a remote environment is first updated, _and then_ pushed.
If the push fails $FLOX_ENV will point to the updated but unpushed out_link of the inner environment.

RemoteEnvironment now refers to a separate `out_link`,
which is a symlink to the store path of the inner managed environment.
When the push succeeds, that link is updated to point to the new store path.